Move test_util to common_test at reasonable levels
