+++ crates/modelardb_common_test/Cargo.toml
@@ -13,16 +13,11 @@
 # limitations under the License.
 
 [package]
-name = "modelardb_compression"
+name = "modelardb_common_test"
 version = "0.1.0"
 license = "Apache-2.0"
 edition = "2021"
 authors = ["Soeren Kejser Jensen <devel@kejserjensen.dk>"]
 
 [dependencies]
-arrow.workspace = true
-modelardb_common = { path = "../modelardb_common" }
-
-[dev-dependencies]
-proptest.workspace = true
-modelardb_common_test = { path = "../modelardb_common_test" }
+rand.workspace = true
