+++ crates/modelardb_server/src/query/mod.rs
@@ -554,14 +554,14 @@ mod tests {
     use datafusion::prelude::Expr;
     use modelardb_common::types::Timestamp;
 
-    use crate::metadata::test_util;
+    use crate::common_test;
 
     const TIMESTAMP_PREDICATE_VALUE: Timestamp = 37;
 
     // Tests for rewrite_and_combine_filters().
     #[test]
     fn test_rewrite_empty_vec() {
-        let schema = test_util::model_table_metadata().schema;
+        let schema = common_test::model_table_metadata().schema;
         let (parquet_filter, grid_filter) = rewrite_and_combine_filters(&schema, &[]);
         assert!(parquet_filter.is_none());
         assert!(grid_filter.is_none());
@@ -570,7 +570,7 @@ mod tests {
     #[test]
     fn test_rewrite_greater_than_timestamp() {
         let filters = new_timestamp_filters(Operator::Gt);
-        let schema = test_util::model_table_metadata().schema;
+        let schema = common_test::model_table_metadata().schema;
         let (parquet_filter, grid_filter) = rewrite_and_combine_filters(&schema, &filters);
 
         assert_binary_expr(
@@ -591,7 +591,7 @@ mod tests {
     #[test]
     fn test_rewrite_greater_than_or_equal_timestamp() {
         let filters = new_timestamp_filters(Operator::GtEq);
-        let schema = test_util::model_table_metadata().schema;
+        let schema = common_test::model_table_metadata().schema;
         let (parquet_filter, grid_filter) = rewrite_and_combine_filters(&schema, &filters);
 
         assert_binary_expr(
@@ -612,7 +612,7 @@ mod tests {
     #[test]
     fn test_rewrite_less_than_timestamp() {
         let filters = new_timestamp_filters(Operator::Lt);
-        let schema = test_util::model_table_metadata().schema;
+        let schema = common_test::model_table_metadata().schema;
         let (parquet_filter, grid_filter) = rewrite_and_combine_filters(&schema, &filters);
 
         assert_binary_expr(
@@ -633,7 +633,7 @@ mod tests {
     #[test]
     fn test_rewrite_less_than_or_equal_timestamp() {
         let filters = new_timestamp_filters(Operator::LtEq);
-        let schema = test_util::model_table_metadata().schema;
+        let schema = common_test::model_table_metadata().schema;
         let (parquet_filter, grid_filter) = rewrite_and_combine_filters(&schema, &filters);
 
         assert_binary_expr(
@@ -654,7 +654,7 @@ mod tests {
     #[test]
     fn test_rewrite_equal_timestamp() {
         let filters = new_timestamp_filters(Operator::Eq);
-        let schema = test_util::model_table_metadata().schema;
+        let schema = common_test::model_table_metadata().schema;
         let (parquet_filter, grid_filter) = rewrite_and_combine_filters(&schema, &filters);
 
         if let Expr::BinaryExpr(BinaryExpr { left, op, right }) = parquet_filter.unwrap() {
