+++ crates/modelardb_server/src/storage/compressed_data_buffer.rs
@@ -118,13 +118,13 @@ impl CompressedDataBuffer {
 mod tests {
     use super::*;
 
-    use crate::storage::test_util;
+    use crate::common_test;
 
     #[test]
     fn test_can_append_valid_compressed_segments() {
         let mut compressed_data_buffer = CompressedDataBuffer::new();
         compressed_data_buffer
-            .append_compressed_segments(test_util::compressed_segments_record_batch());
+            .append_compressed_segments(common_test::compressed_segments_record_batch());
 
         assert_eq!(compressed_data_buffer.compressed_segments.len(), 1)
     }
@@ -133,7 +133,7 @@ mod tests {
     fn test_compressed_data_buffer_size_updated_when_appending() {
         let mut compressed_data_buffer = CompressedDataBuffer::new();
         compressed_data_buffer
-            .append_compressed_segments(test_util::compressed_segments_record_batch());
+            .append_compressed_segments(common_test::compressed_segments_record_batch());
 
         assert!(compressed_data_buffer.size_in_bytes > 0);
     }
@@ -141,7 +141,7 @@ mod tests {
     #[test]
     fn test_can_save_compressed_data_buffer_to_apache_parquet() {
         let mut compressed_data_buffer = CompressedDataBuffer::new();
-        let segment = test_util::compressed_segments_record_batch();
+        let segment = common_test::compressed_segments_record_batch();
         compressed_data_buffer.append_compressed_segments(segment.clone());
 
         let temp_dir = tempfile::tempdir().unwrap();
@@ -167,11 +167,11 @@ mod tests {
 
     #[test]
     fn test_get_size_of_compressed_data_buffer() {
-        let compressed_data_buffer = test_util::compressed_segments_record_batch();
+        let compressed_data_buffer = common_test::compressed_segments_record_batch();
 
         assert_eq!(
             CompressedDataBuffer::size_of_compressed_segments(&compressed_data_buffer),
-            test_util::COMPRESSED_SEGMENTS_SIZE,
+            common_test::COMPRESSED_SEGMENTS_SIZE,
         );
     }
 }
